Pressing Enter after death will start new game
Separated PlayerController from Player object. This allows
actual delettion of the player, rather than having to
just temporarily disable it.

This means that our trail no longer remains after a restart,
and we're able to keep listening for key inputs when the
game is paused (since our script isn't set to inactive).